Created TemplatingEngine
`TemplatingEngine` has been defined as an interface. PHP Printer will provide
out of the box an implementation using an in-house simple templating language
named Simpla, but contributions are welome to provide implementations for
third party ones (such as Twig, Blade, Smarty, etc).

It defines a `render` method which will:

1. get a template using the given name
2. replace the template's placeholders by the values from the given parameters
3. return the resulting string

If no templates are found, it will throw a `TemplateNotFound` exception.