Changed the definition of public API
Instead of using the `@api` annotation to mark the public API, we're going to
consider every classes/interfaces/traits defined in the root namespace as part
of it.

So for example `Memio\PhpPrinter\TemplatingEngine` would be part of the
public API. If we declare a `Memio\PhpPrinter\Simpla\SimplaTemplatingEngine`
class then it won't be part of the public API, because it will be one level
deeper than the root namespace.