Created SimplaTemplateCollection
To allow our incoming SimplaTemplatingEngine implementation to find a template
by a given name, we've created this collection that will be injected into it.

This collection can register a template path or the given name.

In order to be able to overwrite an already given template, it accepts a
priority parameter: the path registered with the highest priority will
be used. If a template is registered twice with the same priority, the latest
addition will be used.