Created ComposedSimplaRule
This implementation of SimplaRule applies all the registered SimplaRules
to the given template, using the given parameters.

Each SimplaRule will returned a transformed template that will be then passed
to the next one.

SimplaRules are registered in ComposedSimplaRule with a priority that allows us
to control in which order they are executed: the highest the priority, the
sooner it is applied.

If two SimplaRules are registered with the same priority, then the later one
will be used.